requirements.txt
PyPDF2 
pdf2image 
opencv-python 
pyzbar
numpy
flask
python-dotenv
selenium